requirements.txt
numpy==2.3.3
scipy==1.16.2
soundfile==0.13.1
praat-textgrids==1.4.0
requests==2.32.5
tqdm==4.67.1
matplotlib==3.10.6
